+++ src/main/java/com/project/config/IPConfig.java
@@ -5,6 +5,8 @@ import javax.servlet.http.HttpSession;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
+import com.project.common.constant.ConstantConfig;
+
 //진짜 단순하게 ip를 넣고 뺴고 끝
 public class IPConfig {
