+++ src/main/java/com/project/controller/BoardController.java
@@ -19,9 +19,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.project.common.MyResponseEntity;
-import com.project.dto.BoardDTO;
-import com.project.dto.BoardDetailDTO;
 import com.project.dto.MyResponseEntityDTO;
+import com.project.dto.board.BoardDTO;
+import com.project.dto.board.BoardDetailDTO;
 import com.project.dto.MainDTO;
 import com.project.service.BoardService;
