+++ src/main/java/com/project/dto/board/BoardDetailDTO.java
@@ -1,36 +1,33 @@
-package com.project.dto;
+
+package com.project.dto.board;
 
 import java.util.Date;
+import java.util.List;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
-import org.springframework.format.annotation.DateTimeFormat;
+import com.project.dto.reply.ReplyDTO;
 
-public class BoardDTO {
+public class BoardDetailDTO {
 
-	@NotNull
-    private Integer boardNum;
 	@NotBlank
-    private String catDomain;
+	private Integer boardNum;
 	@NotBlank
-    private String creator;
+	private String catDomain;
 	@NotBlank
-    private String ttl;
+	private String creator;
 	@NotBlank
-    private String cn;
-	@NotNull
-    private Integer viewCnt;
-	@NotNull
-    private Integer rplCnt;
-    @NotBlank
-    @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSX")
-    private Date crtTm;
-    @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSX")
-    private Date upTm;
-    private String pw;
+	private String ttl;
+	@NotBlank
+	private String cn;
+	private Integer viewCnt;
+	private Integer rplCnt;
+	private Date crtTm;
+	private Date upTm;
+	private String upNm;
+	private List<ReplyDTO>list;
 
-    public Integer getBoardNum() {
+	public Integer getBoardNum() {
 		return boardNum;
 	}
 	public String getCatDomain() {
@@ -57,8 +54,11 @@ public class BoardDTO {
 	public Date getUpTm() {
 		return upTm;
 	}
-	public String getPw() {
-		return pw;
+	public String getUpNm() {
+		return upNm;
+	}
+	public List<ReplyDTO> getList() {
+		return list;
 	}
 	public void setBoardNum(Integer boardNum) {
 		this.boardNum = boardNum;
@@ -87,17 +87,17 @@ public class BoardDTO {
 	public void setUpTm(Date upTm) {
 		this.upTm = upTm;
 	}
-	public void setPw(String pw) {
-		this.pw = pw;
+	public void setUpNm(String upNm) {
+		this.upNm = upNm;
 	}
-
-	public BoardDTO() {
+	public void setList(List<ReplyDTO> list) {
+		this.list = list;
 	}
 
 	@Override
 	public String toString() {
 		StringBuilder builder = new StringBuilder();
-		builder.append("BoardDTO [boardNum=");
+		builder.append("BoardDetailDTO [boardNum=");
 		builder.append(boardNum);
 		builder.append(", catDomain=");
 		builder.append(catDomain);
@@ -115,8 +115,10 @@ public class BoardDTO {
 		builder.append(crtTm);
 		builder.append(", upTm=");
 		builder.append(upTm);
-		builder.append(", pw=");
-		builder.append(pw);
+		builder.append(", upNm=");
+		builder.append(upNm);
+		builder.append(", list=");
+		builder.append(list);
 		builder.append("]");
 		return builder.toString();
 	}
