+++ src/main/java/com/project/dto/category/CategoryDTO.java
@@ -1,4 +1,4 @@
-package com.project.dto;
+package com.project.dto.category;
 
 import java.util.Date;
