+++ src/main/java/com/project/dto/mb/InsertUserRoleDTO.java
@@ -1,4 +1,4 @@
-package com.project.dto;
+package com.project.dto.mb;
 
 public class InsertUserRoleDTO {
