+++ src/main/java/com/project/dto/mb/MbDTO.java
@@ -1,4 +1,4 @@
-package com.project.dto;
+package com.project.dto.mb;
 
 import java.util.Date;
