+++ src/main/java/com/project/dto/mb/SearchInfoDTO.java
@@ -1,4 +1,4 @@
-package com.project.dto;
+package com.project.dto.mb;
 
 public class SearchInfoDTO {
