+++ src/main/java/com/project/dto/reply/RequestReplyDTO.java
@@ -1,30 +1,32 @@
-package com.project.dto;
+package com.project.dto.reply;
 
 import java.util.List;
 
-public class RequestBoardDTO {
-	private List<BoardDTO> list;
+public class RequestReplyDTO {
+	private List<ReplyDTO> list;
 
-	public RequestBoardDTO() {
-		super();
+	public RequestReplyDTO() {
 	}
 
-	public List<BoardDTO> getList() {
+	public List<ReplyDTO> getList() {
 		return list;
 	}
 
-	public void setList(List<BoardDTO> list) {
+	public void setList(List<ReplyDTO> list) {
 		this.list = list;
 	}
 
 	@Override
 	public String toString() {
 		StringBuilder builder = new StringBuilder();
-		builder.append("RequestBoardDTO [list=");
+		builder.append("RequestReplyDTO [list=");
 		builder.append(list);
 		builder.append("]");
 		return builder.toString();
 	}
+
+
+
 
 
 }
